requirements.txt
streamlit
python-dotenv
langchain==0.3.27
langchain-openai==0.3.35
langchain-community==0.3.17
PyPDF2
faiss-cpu
openai
sentence-transformers
gradio